requirements.txt
pyyaml
Mastodon.py==1.5.2